+++ src/components/GoogleMap.jsx
@@ -27,6 +27,7 @@ export const GoogleMap = ({appointments}) => {
         <Map
           style={{width: '80vw', height: '50vh'}}
           defaultCenter={locations[0] ? locations[0].location : {lat: 34.0549, lng: -118.2426}}
+          mapId='[MAP KEY HERE]'
           defaultZoom={11}
           gestureHandling={'greedy'}
           disableDefaultUI={false}
